feat: refactor code to reuse
