+++ src/data/repositories/PaginatedSurveyD2Repository.ts
@@ -7,32 +7,16 @@ import { Id } from "../../domain/entities/Ref";
 import { apiToFuture, FutureData } from "../api-futures";
 import _ from "../../domain/entities/generic/Collection";
 
-import { Survey, SURVEY_FORM_TYPES, SURVEY_STATUSES } from "../../domain/entities/Survey";
+import { Survey, SURVEY_FORM_TYPES } from "../../domain/entities/Survey";
 import { PaginatedSurveyRepository } from "../../domain/repositories/PaginatedSurveyRepository";
 import { PaginatedReponse } from "../../domain/entities/TablePagination";
-import {
-    getParentDataElementForProgram,
-    getSurveyNameBySurveyFormType,
-    isTrackerProgram,
-} from "../utils/surveyProgramHelper";
-import {
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRF,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRL,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_PIS,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SRL,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SSTF,
-    PREVELANCE_SURVEY_NAME_DATAELEMENT_ID,
-    SURVEY_ID_DATAELEMENT_ID,
-    SURVEY_ID_FACILITY_LEVEL_DATAELEMENT_ID,
-    SURVEY_ID_PATIENT_DATAELEMENT_ID,
-    SURVEY_NAME_DATAELEMENT_ID,
-    SURVEY_PATIENT_CODE_DATAELEMENT_ID,
-    WARD_ID_DATAELEMENT_ID,
-} from "./SurveyFormD2Repository";
-import { D2TrackerEvent } from "@eyeseetea/d2-api/api/trackerEvents";
+import { getParentDataElementForProgram, isTrackerProgram } from "../utils/surveyProgramHelper";
+import { WARD_ID_DATAELEMENT_ID } from "../entities/D2Survey";
+
+import { SurveyD2Repository } from "./SurveyFormD2Repository";
 
 export class PaginatedSurveyD2Repository implements PaginatedSurveyRepository {
-    constructor(private api: D2Api) {}
+    constructor(private api: D2Api, private surveyFormRepository: SurveyD2Repository) {}
 
     getSurveys(
         surveyFormType: SURVEY_FORM_TYPES,
@@ -61,6 +45,7 @@ export class PaginatedSurveyD2Repository implements PaginatedSurveyRepository {
               );
     }
 
+    //Currently tracker programs are only in Prevelance module
     getTrackerProgramSurveys(
         surveyFormType: SURVEY_FORM_TYPES,
         programId: Id,
@@ -85,44 +70,10 @@ export class PaginatedSurveyD2Repository implements PaginatedSurveyRepository {
                 filter: `${filterParentDEId}:eq:${parentId}`,
             })
         ).flatMap(trackedEntities => {
-            const surveys = trackedEntities.instances.map(trackedEntity => {
-                const parentPrevalenceSurveyId =
-                    trackedEntity.attributes?.find(
-                        attribute =>
-                            attribute.attribute === SURVEY_ID_FACILITY_LEVEL_DATAELEMENT_ID ||
-                            attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SSTF ||
-                            attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRL ||
-                            attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_PIS ||
-                            attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SRL ||
-                            attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRF
-                    )?.value ?? "";
-
-                return this.getSurveyNameFromId(parentPrevalenceSurveyId, "Prevalence").map(
-                    parentppsSurveyName => {
-                        const survey: Survey = {
-                            id: trackedEntity.trackedEntity ?? "",
-                            name: trackedEntity.trackedEntity ?? "",
-                            rootSurvey: {
-                                id: parentPrevalenceSurveyId ?? "",
-                                name: parentppsSurveyName,
-                                surveyType: "",
-                            },
-                            startDate: trackedEntity.createdAt
-                                ? new Date(trackedEntity.createdAt)
-                                : undefined,
-                            status: "ACTIVE",
-                            assignedOrgUnit: {
-                                id: trackedEntity.orgUnit ?? "",
-                                name: "",
-                            },
-                            surveyType: "",
-                            parentWardRegisterId: undefined,
-                            surveyFormType: surveyFormType,
-                        };
-                        return survey;
-                    }
-                );
-            });
+            const surveys = this.surveyFormRepository.mapTrackedEntityToSurvey(
+                trackedEntities,
+                surveyFormType
+            );
 
             return Future.sequential(surveys).map(surveys => {
                 return {
@@ -160,58 +111,11 @@ export class PaginatedSurveyD2Repository implements PaginatedSurveyRepository {
         ).flatMap(response => {
             const events = response.instances;
 
-            const surveys = events.map(event => {
-                let startDateString,
-                    parentPPSSurveyId = "",
-                    parentWardRegisterId = "",
-                    patientCode = "";
-
-                event.dataValues.forEach(dv => {
-                    if (
-                        dv.dataElement === SURVEY_ID_DATAELEMENT_ID ||
-                        dv.dataElement === SURVEY_ID_PATIENT_DATAELEMENT_ID
-                    )
-                        parentPPSSurveyId = dv.value;
-
-                    if (dv.dataElement === WARD_ID_DATAELEMENT_ID) parentWardRegisterId = dv.value;
-
-                    if (dv.dataElement === SURVEY_PATIENT_CODE_DATAELEMENT_ID)
-                        patientCode = dv.value;
-                });
-
-                const startDate = startDateString ? new Date(startDateString) : undefined;
-
-                return this.getSurveyNameFromId(parentPPSSurveyId, "PPS").map(
-                    parentppsSurveyName => {
-                        const survey: Survey = {
-                            id: event.event,
-                            name: getSurveyNameBySurveyFormType(surveyFormType, {
-                                eventId: event.event,
-                                surveyName: "",
-                                orgUnitName: event.orgUnitName,
-                                hospitalCode: "",
-                                wardCode: "",
-                                patientCode,
-                            }),
-                            rootSurvey: {
-                                id: parentPPSSurveyId,
-                                name: parentppsSurveyName,
-                                surveyType: "",
-                            },
-                            startDate: startDate,
-                            status:
-                                event.status === "COMPLETED"
-                                    ? ("COMPLETED" as SURVEY_STATUSES)
-                                    : ("ACTIVE" as SURVEY_STATUSES),
-                            assignedOrgUnit: { id: event.orgUnit, name: event.orgUnitName ?? "" },
-                            surveyType: "",
-                            parentWardRegisterId: parentWardRegisterId,
-                            surveyFormType: surveyFormType,
-                        };
-                        return survey;
-                    }
-                );
-            });
+            const surveys = this.surveyFormRepository.mapEventToSurvey(
+                events,
+                surveyFormType,
+                programId
+            );
 
             return Future.sequential(surveys).map(surveys => {
                 return {
@@ -223,39 +127,6 @@ export class PaginatedSurveyD2Repository implements PaginatedSurveyRepository {
                     objects: surveys,
                 };
             });
-        });
-    }
-
-    getSurveyNameFromId(id: Id, parentSurveyType: "PPS" | "Prevalence"): FutureData<string> {
-        if (id !== "")
-            return this.getEventProgramById(id)
-                .flatMap(survey => {
-                    if (survey) {
-                        if (parentSurveyType === "PPS") {
-                            const ppsSurveyName = survey.dataValues?.find(
-                                dv => dv.dataElement === SURVEY_NAME_DATAELEMENT_ID
-                            )?.value;
-                            return Future.success(ppsSurveyName ?? "");
-                        } else {
-                            const prevalenceSurveyName = survey.dataValues?.find(
-                                dv => dv.dataElement === PREVELANCE_SURVEY_NAME_DATAELEMENT_ID
-                            )?.value;
-                            return Future.success(prevalenceSurveyName ?? "");
-                        }
-                    } else return Future.success("");
-                })
-                .flatMapError(_err => Future.success(""));
-        else return Future.success("");
-    }
-
-    getEventProgramById(eventId: Id): FutureData<D2TrackerEvent | void> {
-        return apiToFuture(
-            this.api.tracker.events.getById(eventId, {
-                fields: { $all: true },
-            })
-        ).flatMap(resp => {
-            if (resp) return Future.success(resp);
-            else return Future.success(undefined);
         });
     }
 }
